removing sample_helpers.py
